+++ src/utils/gpu/intel.rs
@@ -107,12 +107,12 @@ impl GpuImpl for IntelGpu {
         Ok(true)
     }
 
-    fn used_vram(&self) -> Result<usize> {
-        self.drm_used_vram().map(|usage| usage as usize)
+    fn used_vram(&self) -> Result<u64> {
+        self.drm_used_vram().map(|usage| usage as u64)
     }
 
-    fn total_vram(&self) -> Result<usize> {
-        self.drm_total_vram().map(|usage| usage as usize)
+    fn total_vram(&self) -> Result<u64> {
+        self.drm_total_vram().map(|usage| usage as u64)
     }
 
     fn temperature(&self) -> Result<f64> {
